Added a new slider for changing the increment size (no application in code yet), added ability to set initial times table value upon starting a new visualization

+++ src/Visualization.java
@@ -2,11 +2,11 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Visualization{
-    private int timesTable;
+    private double timesTable;
     private List<Dot> dotList;
     private int numDots;
 
-    public Visualization(int times, int dots){
+    public Visualization(double times, int dots){
         timesTable = times;
         numDots = dots;
         setDotList();
